Move recent_measurements() and measurements_by_metric() to external.database.measurements.
