+++ components/server/src/external/routes/measurement.py
@@ -9,11 +9,7 @@ import bottle
 from pymongo.database import Database
 
 from database.datamodels import latest_datamodel
-from database.measurements import (
-    measurements_by_metric,
-    insert_new_measurement,
-    latest_measurement,
-)
+from database.measurements import insert_new_measurement, latest_measurement
 from database.reports import latest_reports
 from external.database import sessions
 from model.data import SourceData
@@ -21,7 +17,7 @@ from model.measurement import Measurement
 from model.metric import Metric
 from server_utilities.type import MetricId, SourceId
 
-from ..database.measurements import count_measurements
+from ..database.measurements import count_measurements, measurements_by_metric
 from ..utils.functions import report_date_time
 
 from .plugins.auth_plugin import EDIT_ENTITY_PERMISSION
