+++ components/server/src/external/routes/reports_overview.py
@@ -4,11 +4,11 @@ import bottle
 from pymongo.database import Database
 
 from database.datamodels import latest_datamodel
-from database.measurements import recent_measurements
 from database.reports import insert_new_reports_overview, latest_reports, latest_reports_overview
 from model.transformations import hide_credentials
 
 from ..database import sessions
+from ..database.measurements import recent_measurements
 from ..utils.functions import report_date_time, sanitize_html
 
 from .plugins.auth_plugin import EDIT_REPORT_PERMISSION
